pgvectorscale-rag-solution/requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
PyPDF2
langchain
langchain-huggingface
huggingface_hub
pgvector
SQLAlchemy
psycopg2-binary
pickle-mixin
pinecone-client
python-dotenv
Jinja2
pypdf2
langchain_community
faiss-cpu
transformers
torch
sentence-transformers
reportlab
pydantic
instructor
anthropic



